AuditAgent/__init__.py
import azure.functions as func
import logging

def main(req: func.HttpRequest) -> func.HttpResponse:
    logging.info('=== Audit Agent Started ===')
    logging.info('Immutable audit logging would happen here')
    return func.HttpResponse("Audit Agent executed", status_code=200)
