requirements.txt
azure-functions==1.18.0
pymongo==4.6.1
pika==1.3.2

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
mongomock==4.1.2
python-dotenv==1.0.1
